Started entity component refactory
